Refactor course addition forms
