+++ src/pages/dashboard/dashboard.js
@@ -1,54 +1,7 @@
 import 'normalize.css/normalize.css';
 import '../../styles/index.scss';
 import { initRoleSelection, bindRoleSelection } from '../../utils/role-modal';
-
-// Show the appropriate form for a user's role
-function initAddCourseForm() {
-  const role = localStorage.role;
-
-  if (role === 'student') {
-    $('.professor-form').remove();
-  }
-  else if (role === 'professor') {
-    $('.student-form').remove();
-  }
-};
-
-// Add course to list in sidebar
-function addCourseToSideBar(course) {
-  $('.sidebar .courses').append(`
-    <div class="course-info">
-      <h5>Course Name: ${course.courseName}</h5>
-      <p>Professor: ${course.professor}</p>
-      <p>Join Code: ${course.joinCode}</p>
-    </div>
-  `);
-};
-
-function getAllCourses() {
-  const user = netlifyIdentity.currentUser();
-
-  fetch('/.netlify/functions/get-courses', {
-    method: "POST",
-    body: JSON.stringify({ email: user.email })
-  })
-  .then(response => {
-    if (!response.ok) {
-      return response
-        .text()
-        .then(err => {throw(err)});
-    }
-
-    response.text().then(function(result) {
-      const { courses } = JSON.parse(result);
-
-      if (courses) {
-        courses.map(course => addCourseToSideBar(course));
-      };
-    });
-  })
-  .catch(err => console.error(err));
-};
+import { initAddCourseForm, addCourseToSideBar, getAllCourses, bindAddCourseButton } from '../../utils/course-utils';
 
 // Redirect the user to the home page if not logged in
 $(window).on('load', () => {
@@ -82,58 +35,8 @@ $('document').ready(function() {
   // Bind on-click handler for selecting a role in the modal
   bindRoleSelection();
 
-  // Takes the information from the form and submits it to add-course Netlify function
-  $('.add-course-button').on('click', function() {
-    const user = netlifyIdentity.currentUser();
-    const role = localStorage.role;
-    const form = $(`#add-course-form .${role}-form`);
-    
-    let courseInfo = {};
-    
-    if (role === 'student') {
-
-      const joinCode = form.find('input').get(0).value;
-
-      courseInfo = {
-        email: user.email,
-        role: role,
-        joinCode: joinCode
-      };
-    }
-    else if (role === 'professor') {
-
-      const school = form.find('input').get(0).value;
-      const course = form.find('input').get(1).value;
-
-      courseInfo = {
-        name: user.user_metadata.full_name,
-        email: user.email,
-        role: role,
-        course: course,
-        school: school
-      };
-    }
-
-    fetch('/.netlify/functions/add-course', {
-      method: "POST",
-      body: JSON.stringify(courseInfo)
-    })
-    .then(response => {
-      if (!response.ok) {
-        return response
-          .text()
-          .then(err => {throw(err)});
-      }
-
-      response.text().then(function(result) {
-        const course = JSON.parse(result);
-        addCourseToSideBar(course);
-      });
-
-      $('#add-course-modal').modal('hide');
-    })
-    .catch(err => console.error(err));
-  });
+  // Bind on-click handler for adding a course
+  bindAddCourseButton();
 });
